Updated user stories for saving and loading game states. Updated the temporary game map for ease of testing.
Wrote unit tests for JsonReader and JsonWriter. Implemented Writable interface through classes JsonReader and JsonWriter. Added toJson() methods for Block, PowerUp, Character, and Game classes. The character field of Game is no longer final to allow for loading of saved games.

Added class comments to GameTerminal and Main. Changed drawTime() to drawHUD() in GameTerminal to include instructions to save the game. Added drawMainMenu() and handleMenuInputs() to GameTerminal to illustrate and handle the options of starting a new game or loading a saved game. Added loadGame() and saveGame() to GameTerminal to perform saving and loading functions.